Storage added

+++ src/app/components/menu/list/list.component.ts
@@ -1,10 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { FormGroup, FormArray, FormBuilder, FormControl, Validators } from '@angular/forms';
 import { ExerciseInterface, Exercise } from 'src/app/services/exercises/exercises.models';
 import { ExercisesService } from 'src/app/services/exercises/exercises.service';
-import { MenuController } from '@ionic/angular';
+import { MenuController, IonReorderGroup } from '@ionic/angular';
 import { EventsService } from 'src/app/services/events/events.service';
-import { Keyboard } from '@ionic-native/keyboard/ngx';
+import { ClockerService } from 'src/app/services/clocker/clocker.service';
 
 @Component({
   selector: 'app-list',
@@ -12,16 +12,22 @@ import { Keyboard } from '@ionic-native/keyboard/ngx';
   styleUrls: ['./list.component.scss'],
 })
 export class ListComponent implements OnInit {
+  //Component Variables 
+  @ViewChild (IonReorderGroup,{static:false}) reorderGroup: IonReorderGroup;
+  
+  //Class Variable
+  isReordering:boolean = false
+
   constructor(
     public exercisesService:ExercisesService,
     public menu:MenuController,
     public eventsService:EventsService,
-    public keyboard:Keyboard
+    public clockerService:ClockerService
   ) {
    }
 
   ngOnInit() {
-
+    
   }
 
   async onPlay(){
@@ -36,9 +42,20 @@ export class ListComponent implements OnInit {
     return this.exercisesService.hasExercises() ? this.exercisesService.exercises.length + ' exercise(s)' : ''
   }
   onDeleteExercise(exercise:Exercise){
-    console.log(exercise);
-    
-    this.exercisesService.remove(exercise)
+    if(!this.clockerService.intervalTimer.isRunning){
+      this.exercisesService.remove(exercise)
+    }
+  }
+  onToggleReorder(){
+    this.isReordering = !this.isReordering
+  }
+  disabledReorder(){
+    return !this.exercisesService.hasExercises(1)
+  }
+  onReorder(ev: any){
+    if(ev.detail){
+      ev.detail.complete(this.exercisesService.exercises);
+    }
   }
 }
